requirements.txt
Package               Version
--------------------- -----------
absl-py               2.1.0
appnope               0.1.4
asttokens             3.0.0
attrs                 25.2.0
bcrypt                4.3.0
bidict                0.23.1
blinker               1.9.0
cffi                  1.17.1
click                 8.1.8
comm                  0.2.2
contourpy             1.3.0
cycler                0.12.1
debugpy               1.8.13
decorator             5.2.1
dnspython             2.7.0
email_validator       2.2.0
exceptiongroup        1.2.2
executing             2.2.0
Flask                 3.1.0
Flask-Bcrypt          1.0.1
flask-cors            5.0.1
Flask-Login           0.6.3
Flask-Mail            0.10.0
Flask-PyMongo         3.0.1
Flask-SocketIO        5.5.1
Flask-WTF             1.2.2
flatbuffers           25.2.10
fonttools             4.56.0
h11                   0.14.0
idna                  3.10
importlib_metadata    8.6.1
importlib_resources   6.5.2
ipykernel             6.29.5
ipython               8.18.1
itsdangerous          2.2.0
jax                   0.4.30
jaxlib                0.4.30
jedi                  0.19.2
Jinja2                3.1.6
joblib                1.4.2
jupyter_client        8.6.3
jupyter_core          5.7.2
kiwisolver            1.4.7
MarkupSafe            3.0.2
matplotlib            3.9.4
matplotlib-inline     0.1.7
mediapipe             0.10.21
ml_dtypes             0.5.1
nest-asyncio          1.6.0
numpy                 1.26.4
opencv-contrib-python 4.11.0.86
opencv-python         4.11.0.86
opt_einsum            3.4.0
packaging             24.2
pandas                2.2.3
parso                 0.8.4
pexpect               4.9.0
pillow                11.1.0
pip                   24.3.1
platformdirs          4.3.6
prompt_toolkit        3.0.50
protobuf              4.25.6
psutil                7.0.0
ptyprocess            0.7.0
pure_eval             0.2.3
pycparser             2.22
Pygments              2.19.1
pymongo               4.11.2
pyparsing             3.2.1
python-dateutil       2.9.0.post0
python-dotenv         1.0.1
python-engineio       4.11.2
python-socketio       5.12.1
pytz                  2025.1
pyzmq                 26.3.0
scikit-learn          1.6.1
scipy                 1.13.1
sentencepiece         0.2.0
setuptools            75.6.0
simple-websocket      1.1.0
six                   1.17.0
sounddevice           0.5.1
stack-data            0.6.3
threadpoolctl         3.5.0
tornado               6.4.2
traitlets             5.14.3
typing_extensions     4.12.2
tzdata                2025.1
wcwidth               0.2.13
Werkzeug              3.1.3
wsproto               1.2.0
WTForms               3.2.1
zipp                  3.21.0
